Close and Exit buttons fix

+++ SistemaHotelaria/Model/Quarto.cs
@@ -16,13 +16,13 @@ namespace SistemaHotelaria.Model
 
         private int id;
         private String categoria;
-        private float valor;
+        private decimal valor;
         private String numero;
         private String status;
 
         public int Id { get => id; set => id = value; }
         public string Categoria { get => categoria; set => categoria = value; }
-        public float Valor { get => valor; set => valor = value; }
+        public decimal Valor { get => valor; set => valor = value; }
         public string Numero { get => numero; set => numero = value; }
         public string Status { get => status; set => status = value; }
     }
